requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
PyYAML>=5.4.0
tqdm>=4.60.0
